feat(nx-dev): move all the querying logic to edge function
