+++ nx-dev/util-ai/README.md
@@ -1,0 +1,11 @@
+# nx-dev-util-ai
+
+This library was generated with [Nx](https://nx.dev).
+
+## Building
+
+Run `nx build nx-dev-util-ai` to build the library.
+
+## Running unit tests
+
+Run `nx test nx-dev-util-ai` to execute the unit tests via [Jest](https://jestjs.io).
