+++ nx-dev/util-ai/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@nx/nx-dev/util-ai",
+  "version": "0.0.1",
+  "dependencies": {
+    "tslib": "^2.3.0",
+    "openai": "~3.3.0"
+  },
+  "type": "commonjs",
+  "main": "./src/index.js",
+  "typings": "./src/index.d.ts"
+}
